Day 8 commit- Offline Exercise

+++ src/main/scala/Garage/Part.scala
@@ -3,10 +3,10 @@ package Garage
 /**
   * Created by Profile on 13/06/2017.
   */
-class Part(id1:Int,title1:String,price1:Double,hoursToRepair:Int) {
+class Part(id1:Int,title1:String,price1:Int,hoursToRepair:Int) {
   def id:Int=id1
   def title:String=title1
-  def price:Double=price1
+  def price:Int=price1
   def repairHours:Int=hoursToRepair
   var isBroken:Boolean=false
